+++ viewtopic.php
@@ -345,7 +345,7 @@ $sql = "
 	WHERE p.topic_id = $topic_id
 		$limit_posts_time
 	GROUP BY p.post_id
-	ORDER BY p.post_time $post_order
+	ORDER BY p.post_pinned DESC, p.post_time $post_order
 	LIMIT $start, $posts_per_page
 ";
 
@@ -594,7 +594,7 @@ for ($i = 0; $i < $total_posts; $i++) {
     }
 
     // Buttons
-    $pm_btn = $profile_btn = $delpost_btn = $edit_btn = $ip_btn = $quote_btn = '';
+    $pm_btn = $profile_btn = $delpost_btn = $edit_btn = $ip_btn = $quote_btn = $pin_btn = '';
 
     if (!$poster_guest) {
         $profile_btn = true;
@@ -604,6 +604,7 @@ for ($i = 0; $i < $total_posts; $i++) {
     if (!$poster_bot) {
         $quote_btn = ($is_auth['auth_reply'] && !$locked);
         $edit_btn = (($userdata['user_id'] == $poster_id && $is_auth['auth_edit']) || $is_auth['auth_mod']);
+        $pin_btn = ($post_id != $t_data['topic_first_post_id']) && (($userdata['user_id'] == $poster_id) || $is_auth['auth_mod']);
         $ip_btn = ($is_auth['auth_mod'] || IS_MOD);
     }
     $delpost_btn = ($post_id != $t_data['topic_first_post_id'] && ($is_auth['auth_mod'] || ($userdata['user_id'] == $poster_id && $is_auth['auth_delete'] && $t_data['topic_last_post_id'] == $post_id && $postrow[$i]['post_time'] + 3600 * 3 > TIMENOW)));
@@ -690,6 +691,7 @@ for ($i = 0; $i < $total_posts; $i++) {
         'POSTER_AUTHOR' => ($poster_id == $t_data['topic_poster']),
         'POSTER_GENDER' => !$poster_guest ? genderImage((int)$postrow[$i]['user_gender']) : '',
         'POSTED_AFTER' => $prev_post_time ? delta_time($postrow[$i]['post_time'], $prev_post_time) : '',
+        'IS_PINNED' => $postrow[$i]['post_pinned'],
         'IS_UNREAD' => is_unread($postrow[$i]['post_time'], $topic_id, $forum_id),
         'IS_FIRST_POST' => (!$start && $is_first_post),
         'MOD_CHECKBOX' => ($moderation && ($start || defined('SPLIT_FORM_START'))),
@@ -705,6 +707,7 @@ for ($i = 0; $i < $total_posts; $i++) {
 
         'QUOTE' => $quote_btn,
         'EDIT' => $edit_btn,
+        'PIN' => $pin_btn,
         'DELETE' => $delpost_btn,
         'IP' => $ip_btn,